Updated accelerometer to write only changed values
Noticed accelerometer values were being read faster than the
accelerometer was updating/changing them. In other words the pi
was reading the same unchanged data values multiple times, and
instead of writing that value once it wrote it as many times as
it read it. Changed the script to check for unchanged values, this
works because even at rest the values change a little. Updated the
number of entries per log file, as with less writes the entries
will be fewer