Refactored logging scripts to overwrite instead of append.
Logging scripts were previously appending, so when the circular
log came back to the start,the old data remained and new data
jsut got appended, so much for circular logging. Now fixed.

Also in this commit are an analysis script with Fourier Transform,
and also a shell script which is useful for parsing out a specific
column from the data files, and concatenating them into a single
file, which can then be fed into the analysis script.