Code will now create a new folder when logging script begins
As multiple datasets need to be collected, at least one for each
state, the pi will need to start separate logs for each dataset.
The simplest way is to restart the script by rebooting the pi,
and by creating a new folder each time the logging begins, it
will be much easier to later split the data into different
datasets when they are in separate folders.