Added code to get Lidar variance histograms, worked on state_estimation
Added code to write Lidar variances to file.
Then created copy of pmf as pmfLidar and customised to create
dynamic bin histogram dictionary for lidar variances, which then
wrote to file.

Wrote code in state_estimation for reading in the histogram dicts
and constructing the dictionaries in state_estimation code, next
will be to write the movement models (easy) and then get on with
writing the algorithm in python (transferring from Matlab...)