Simulation code basically complete in Python, model not accurate
Despite the code now basically being complete, the model itself
is not working accurately, which means that the results are not
good in terms of prediction. This is sad but I don't have time
to do any more work on the model, so will have to try and
implement in hardware as it is.