Finally ironed out the bugs in the FFANN Training System, good to go now
The training system to find a suitable FeedForwardArtifialNeuralNetwork
which utilises an Evolutionary Algorithm to optimise the training and
steer the FFANN's toward an optimum solution, has been pretty much
completed now with bugs ironed out. So should be able to use it to now
train suitable FFANNs for the measurement-model aspect of my
State Estimation Hons Proj algorithm.