Modified to read 3 datafiles and get a rough classification for that range
E.g. read Accel15psi 20psi and 25psi data and output the mid psi of 20
but output will be 2000 as the ANNs seem to be able to reach values in
the 1000's more easily.

So then this will roughly classify that psi range of 15->25 as 2000

The point is that then anything in that range can be classified.

So then the next range to train will be 30,35,40 as 3500
then 45 50 55 as 5000 maybe will include an extra psi in that
middle group.