Working now. Also added the ability to choose sensor fusion or not
The latest changes of using empirically-found model probabilities
works okay, it's not the best differentiator but it does give
roughly the right results, which is an improvement on before.

Also added the ability to run just for Lidar, or just for Accel,
or to run using Sensor Fusion with both (which is what has been
used so far), could be useful to compare Sensor Fusion with sans.