Added combined script with Accelerometer real-data logging, plus prediction
Combined the Accelerometer data logging script with the machine-learning
simulation script in order to apply the artificial neural network models
developed to the raw data as it is received from the accelerometer sensor.

The idea will be, initially at least, to do prediction separately for
Accel and Lidar data, then look at the results, then maybe combine results
for the individually predicted results. Later I might try doing the same
as in the simulation script which combines the prediction map for each
data value, but I think it is best to predict separately for each sensor
for now.