Still fixing bugs, getting there.
There were issues with accessing the newpopulation global variable
from within a func then just after from in the main code, so that
it seemed like the appends in the function werent being seen soon
enough in the subsequent main code. Changed that variable to be
inside a class as maybe class has stricter access controls, this
seems to have worked. Now however the input data does not seem to
be getting processed on the 2nd cycle, so need to debug that.