constructFFANN func working, now need to declare all possible inds
So constructFFANN function is working great in and of itself.

But as C requires all variables to be declared in order to be
compiled, I will need to indivually create all the possible
Individuals..... meaning that if I want a population of size 100
I will need to have 100 unique declarations for these Individual
structs. However will actually need 200 in order to keep both
an old pop and a new pop... so next time will create an
additional aux file in which such declarations can be made,
which is kindof annoying as I need to choose an upper limit
for popsize.... I will probably choose 200, therefore will need
400 Individual declarations. This is the pain of C, but also
it will run really quickly once it's ready.