Managed to get the constructFFANN function working correctly.
In order to get it to work, I needed to pre-declare all the
possible Individual structs and also all their possible
Nodes contained in each Individual's input, hidden and output
layers. So that is a lot of pre-declarations, all done in the
individual_aux.c auxiliary code file.

Only tested for calling constructFFANN twice in order to add 2
Individuals to the population, but that now works with unique
weights, biases etc for nodes within each individual. Great!

Next will be to add more individuals.

The downside of this approach however is that the declarations
are hard-coded in individual_aux.c, which is a price you pay in
C becuase you cannot use Lambdas or dynamically-declare structs
on the fly. But as long as you keep within the 200 population
size as hardcoded in individual_aux.c, then you will be fine.
Or you will need to add more declarations if you want larger
population sizes than 200.

Used a python script to write the declaration lines and do stuff
like calling constructNode() for each Node and malloc for
hidden and output layers. The python script makeInd.py is
included in this commit, though I updated it every time I wanted
to add new lines to individual_aux.c so it is not a completely
comprehensive version of makeInd.py