Modified relu method to its proper functionality
Relu was previosuly implemented as essentially absolute.
After checking that the data which is read in is absoluted there,
this means that relu was able to be refactored to its proper
implementation, as any negative values entering a node will be due
to weight values or a bias value, rather than the data itself.

Quick test and the program is still working fine after making this
change.