Copying successfully working. Tournament function in progress.
Copying successfully tested and working.
Able to copy best ANN for elitism for new generation.
Able to select four unique member for tournament.
Able to sort based on LMS in order to pick best 2 ANNs for tourny.

Next is to do the breeding of the chosen pair from the tourny.