Created a python script to generate FFT data batches of datapoints
The script is currently using a sample rate of 2000Hz, so change
as needed.

Matplotlib code is included but currently commented out.

Currently the script is keeping a note of the frequencies where
the largest amplitude occurs and then this list of frequencies
is put through a median calculator and also an avg calculator.
This is working.